Moved all tablets into `fuel-core-interfaces/src/db`.
