+++ fuel-core/src/database.rs
@@ -1,11 +1,5 @@
 use crate::{
-    database::{
-        storage::{
-            FuelBlocks,
-            SealedBlockConsensus,
-        },
-        transactional::DatabaseTransaction,
-    },
+    database::transactional::DatabaseTransaction,
     state::{
         in_memory::memory_store::MemoryStore,
         DataSource,
@@ -25,6 +19,10 @@ use fuel_core_interfaces::{
     common::fuel_storage::{
         StorageAsMut,
         StorageAsRef,
+    },
+    db::{
+        FuelBlocks,
+        SealedBlockConsensus,
     },
     model::{
         BlockHeight,
@@ -63,7 +61,6 @@ use tempfile::TempDir;
 
 // Storages implementation
 // TODO: Move to separate `database/storage` folder, because it is only implementation of storages traits.
-mod balances;
 mod block;
 mod code_root;
 mod coin;
@@ -75,7 +72,6 @@ mod state;
 
 pub mod metadata;
 pub mod resource;
-pub mod storage;
 pub mod transaction;
 pub mod transactional;
 pub mod vm_database;
@@ -88,13 +84,13 @@ pub mod vm_database;
 pub enum Column {
     /// The column id of metadata about the blockchain
     Metadata = 0,
-    /// See [`ContractsRawCode`](fuel_vm::storage::ContractsRawCode)
+    /// See [`ContractsRawCode`](fuel_core_interfaces::db::ContractsRawCode)
     ContractsRawCode = 1,
-    /// See [`ContractsRawCode`](fuel_vm::storage::ContractsRawCode)
+    /// See [`ContractsRawCode`](fuel_core_interfaces::db::ContractsRawCode)
     ContractsInfo = 2,
-    /// See [`ContractsState`](fuel_vm::storage::ContractsState)
+    /// See [`ContractsState`](fuel_core_interfaces::db::ContractsState)
     ContractsState = 3,
-    /// See [`ContractsLatestUtxo`](storage::ContractsLatestUtxo)
+    /// See [`ContractsLatestUtxo`](fuel_core_interfaces::db::ContractsLatestUtxo)
     ContractsLatestUtxo = 4,
     /// See [`ContractsAssets`](fuel_vm::storage::ContractsAssets)
     ContractsAssets = 5,
@@ -108,9 +104,9 @@ pub enum Column {
     TransactionStatus = 9,
     /// The column of the table of all `owner`'s transactions
     TransactionsByOwnerBlockIdx = 10,
-    /// See [`Receipts`](storage::Receipts)
+    /// See [`Receipts`](fuel_core_interfaces::db::Receipts)
     Receipts = 11,
-    /// See [`FuelBlocks`](storage::FuelBlocks)
+    /// See [`FuelBlocks`](fuel_core_interfaces::db::FuelBlocks)
     FuelBlocks = 12,
     /// Maps fuel block id to fuel block hash
     FuelBlockIds = 13,
